backend/python_scripts/requirements.txt
abjad==3.19
alabaster==0.7.16
Babel==2.14.0
black==24.1.1
blinker==1.7.0
boto3==1.34.38
botocore==1.34.38
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docutils==0.20.1
Flask==3.0.2
Flask-Cors==4.0.0
idna==3.6
imagesize==1.4.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
ply==3.11
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
roman==4.1
s3transfer==0.10.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
tomli==2.0.1
typing_extensions==4.9.0
Unidecode==1.3.8
uqbar==0.7.1
urllib3==2.0.7
Werkzeug==3.0.1
